split modal form into separate component files
