+++ src/components/organisms/ModalForm/ModalForm.tsx
@@ -1,43 +1,15 @@
-import {Portal} from '@gorhom/portal';
-import {ReactFormExtendedApi, useForm} from '@tanstack/react-form';
+import {useForm} from '@tanstack/react-form';
 import {Type} from 'arktype/out/type';
-import {BottomMenu, Button, FormInput, SnackBar, useSnackbar} from 'atoms';
-import {FormInputProps} from 'atoms/FormInput/FormInput';
+import {SnackBar, useSnackbar} from 'atoms';
 import {composeTestID} from 'core/helpers';
 import {useBottomMenu, useThemeStyles} from 'core/hooks';
-import {map} from 'lodash';
-import {AnimatedCircleButton, ListItem} from 'molecules';
-import React, {ReactNode, RefObject, memo, useCallback, useState} from 'react';
+import React, {useCallback, useState} from 'react';
 import {TFunction} from 'react-i18next';
-import {
-  Keyboard,
-  Platform,
-  ScrollView,
-  StyleSheet,
-  Text,
-  TextInput,
-  View,
-} from 'react-native';
+import {Platform, View} from 'react-native';
 import {useSafeAreaInsets} from 'react-native-safe-area-context';
-import {FullWindowOverlay} from 'react-native-screens';
-import {isIOS} from 'services/PlatformService';
-import {headerGap, themeStyles} from './styles';
-
-export type FieldConfig = Pick<FormInputProps, 'maxLength' | 'multiline'>;
-type AnyForm<T = any> = ReactFormExtendedApi<
-  T,
-  any,
-  any,
-  any,
-  any,
-  any,
-  any,
-  any,
-  any,
-  any
->;
-
-type AnyFormValues = Record<string, any>;
+import {FieldEditorSheet, FieldSet, Header, SubmitButton} from './components';
+import {themeStyles} from './styles';
+import {AnyFormValues, FieldConfig} from './types';
 
 interface ModalFormProps<T extends AnyFormValues> {
   defaultValues: T;
@@ -161,217 +133,5 @@ export const ModalForm = <T extends AnyFormValues>({
         {...snackBarProps}
       />
     </View>
-  );
-};
-
-const Header = ({
-  title,
-  onBackPress,
-  testID,
-}: {
-  title: ReactNode;
-  onBackPress?: () => void;
-  testID: string;
-}) => {
-  const styles = useThemeStyles(themeStyles);
-
-  const {top: topSafeAreaInset} = useSafeAreaInsets();
-
-  return (
-    <View
-      testID={testID}
-      style={[
-        styles.header,
-        Platform.select({
-          ios: {marginTop: headerGap},
-          android: {marginTop: Math.max(topSafeAreaInset, headerGap * 2)},
-        }),
-      ]}>
-      <AnimatedCircleButton
-        icon={{name: 'chevronMediumLeft'}}
-        testID={composeTestID(testID, 'backButton')}
-        onPress={onBackPress}
-      />
-      <Text testID={composeTestID(testID, 'title')} style={styles.title}>
-        {title}
-      </Text>
-    </View>
-  );
-};
-
-interface SubmitButtonProps {
-  label: string;
-  submitting?: boolean;
-  disabled?: boolean;
-  onPress?: () => void;
-  testID: string;
-}
-
-const SubmitButton = ({
-  onPress: onSubmit,
-  label,
-  submitting,
-  disabled,
-  testID,
-}: SubmitButtonProps) => {
-  return (
-    <Button
-      onPress={onSubmit}
-      theme="primary"
-      testID={testID}
-      text={label}
-      loading={submitting}
-      disabled={disabled}
-    />
-  );
-};
-
-const FieldSetComponent = <T extends AnyForm>({
-  testID,
-  onFieldPress,
-  form,
-  t,
-}: {
-  testID: string;
-  onFieldPress?: (key: Extract<keyof T['state']['values'], string>) => void;
-  form: T;
-  t: TFunction;
-}) => {
-  const styles = useThemeStyles(themeStyles);
-
-  return (
-    <ScrollView
-      showsVerticalScrollIndicator={false}
-      alwaysBounceVertical={false}
-      contentContainerStyle={styles.list}>
-      {map(Object.keys(form.state.values), key => {
-        const valueSelector = (state: typeof form.state) => state.values[key];
-
-        return (
-          <form.Subscribe key={key} selector={valueSelector}>
-            {fieldValue => {
-              return (
-                <ListItem
-                  type="primary"
-                  testID={composeTestID(testID, 'field')}
-                  subtitle={fieldValue}
-                  title={t(`fieldTitles.${key}`)}
-                  onPress={() => onFieldPress?.(key as any)}
-                />
-              );
-            }}
-          </form.Subscribe>
-        );
-      })}
-    </ScrollView>
-  );
-};
-
-const FieldSet = memo(FieldSetComponent) as typeof FieldSetComponent;
-
-const FieldEditorSheet = <T extends string>({
-  testID,
-  fieldName,
-  t,
-  onSubmit,
-  menuProps,
-  onHide,
-  defaultValue,
-  schema,
-  fieldConfig,
-}: {
-  testID: string;
-  fieldName: string;
-  t: TFunction;
-  inputRef?: RefObject<TextInput>;
-  onSubmit?: (value: T) => void;
-  menuProps: ReturnType<typeof useBottomMenu>;
-  onHide?: () => void;
-  defaultValue: T;
-  schema: Type<Record<string, any>>;
-  fieldConfig?: FieldConfig;
-}) => {
-  const {bottom: bottomSafeAreaInset} = useSafeAreaInsets();
-  const form = useForm({
-    defaultValues: {[fieldName]: defaultValue} as Record<string, any>,
-    validators: {
-      onSubmit: schema,
-    },
-    onSubmit: ({value}) => {
-      onSubmit?.(value[fieldName].trim());
-    },
-  });
-
-  const styles = useThemeStyles(themeStyles);
-
-  const fieldBlock = (
-    <form.Field name={fieldName}>
-      {field => {
-        const value = field.state.value;
-        const hasError = !field.state.meta.isValid;
-
-        const errorMessage = hasError
-          ? t(`fieldErrors.${fieldName}`)
-          : undefined;
-
-        return (
-          <View style={styles.fieldContainer}>
-            <Text
-              testID={composeTestID(testID, 'description')}
-              style={styles.fieldDescription}>
-              {t(`fieldDescriptions.${fieldName}`)}
-            </Text>
-            <FormInput
-              ref={menuProps.textInputRef}
-              testID={composeTestID(testID, 'input')}
-              value={value}
-              onChange={field.handleChange}
-              error={hasError}
-              messageText={errorMessage}
-              onSubmitEditing={form.handleSubmit}
-              blurOnSubmit={false}
-              {...fieldConfig}
-            />
-            <SubmitButton
-              onPress={form.handleSubmit}
-              testID={composeTestID(testID, 'submitButton')}
-              label={t('fieldSubmitLabel')}
-              disabled={!value.trim() || hasError}
-            />
-          </View>
-        );
-      }}
-    </form.Field>
-  );
-
-  const menu = (
-    <BottomMenu
-      testID={composeTestID(testID, 'bottomMenu')}
-      withBackdrop
-      initialIndex={0}
-      onBackdropPress={Keyboard.dismiss}
-      adjustIOSKeyboardFrameDrops
-      onHideEnd={onHide}
-      header={{
-        title: t(`fieldTitles.${fieldName}`),
-      }}
-      {...menuProps}>
-      {fieldBlock}
-      <View style={{height: menuProps.keyboardHeight ?? bottomSafeAreaInset}} />
-    </BottomMenu>
-  );
-
-  return (
-    <Portal>
-      {isIOS ? (
-        <FullWindowOverlay>
-          <View pointerEvents="box-none" style={StyleSheet.absoluteFillObject}>
-            {menu}
-          </View>
-        </FullWindowOverlay>
-      ) : (
-        menu
-      )}
-    </Portal>
   );
 };
