+++ src/components/organisms/ModalForm/components/SubmitButton/index.ts
@@ -1,0 +1,1 @@
+export * from './SubmitButton';
